backend/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
aiofiles 
openai 
python-dotenv 